requirements.txt
imaplib2 
streamlit 
python-dotenv
PyPDF2
